+++ src/lib/api/auth/requests/createClientProfile.ts
@@ -1,0 +1,11 @@
+import { ClientProfilePostValue } from "@/lib/schemas";
+import { tokenFetch } from "@/lib/utils";
+
+export default async function createClientProfile(
+   data: ClientProfilePostValue,
+) {
+   await tokenFetch("/profile/clients", {
+      method: "PATCH",
+      body: JSON.stringify(data),
+   });
+}
