+++ src/lib/schemas/auth.schema.ts
@@ -23,7 +23,8 @@ export const checkPasswordSchema = baseAuthSchema
 
 export const nameSchema = baseAuthSchema
    .min(2, "성함을 입력해 주세요.")
-   .max(4, "4자 이내로 입력해 주세요.");
+   .max(4, "4자 이내로 입력해 주세요.")
+   .trim();
 
 export const phoneSchema = baseAuthSchema
    .min(9, "9자 이상 입력해주세요.")
