+++ src/lib/schemas/index.ts
@@ -1,2 +1,3 @@
 // 전체 export 하기
 export * from "./auth.schema";
+export * from "./client.schema";
